pyproject.toml
[tool.poetry]
name = "gh2repo"
version = "0.1.0"
description = ""
authors = ["JoJoJux <80523427+J0J0HA@users.noreply.github.com>"]
license = "GPL"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
hjson = "^3.1.0"
requests = "^2.31.0"
pygithub = "^2.3.0"
dataclasses-json = "^0.6.4"
environs = "^11.0.0"
gitpython = "^3.1.42"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
